+++ RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [typescript v0.2.0] .
 ### Releases
 - [NPM v0.2.0] https://www.npmjs.com/package/@lukehagar/discoursejs/v/0.2.0 - .
+
+## 2024-01-15 15:25:48
+### Changes
+Based on:
+- OpenAPI Doc latest 
+- Speakeasy CLI 1.141.1 (2.233.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.1] .
+### Releases
+- [NPM v0.2.1] https://www.npmjs.com/package/@lukehagar/discoursejs/v/0.2.1 - .